监听积分usdt充值功能实现,部分逻辑优化

+++ ruoyi-system/src/main/java/com/ruoyi/system/api/entity/tronscan/Representative.java
@@ -9,7 +9,7 @@ public class Representative implements Serializable {
 
 
     private static final long serialVersionUID = 1L;
-    private int lastWithDrawTime;
+    private long lastWithDrawTime;
     private int allowance;
     private boolean enabled;
     private String url;
